refactor: remove the vendor.ts/js files
