--- templates/vendor.angular.ts
@@ -1,18 +1,0 @@
-// Snapshot the ~/app.css and the theme
-const application = require("application");
-require("ui/styling/style-scope");
-const appCssContext = require.context("~/", false, /^\.\/app\.(css|scss|less|sass)$/);
-global.registerWebpackModules(appCssContext);
-application.loadAppCss();
-
-require("reflect-metadata");
-require("@angular/platform-browser");
-require("@angular/core");
-require("@angular/common");
-require("@angular/forms");
-require("@angular/http");
-require("@angular/router");
-
-require("nativescript-angular/platform-static");
-require("nativescript-angular/forms");
-require("nativescript-angular/router");
