--- templates/vendor.nativescript.ts
@@ -1,8 +1,0 @@
-// Snapshot the ~/app.css and the theme
-const application = require("application");
-require("ui/styling/style-scope");
-const appCssContext = require.context("~/", false, /^\.\/app\.(css|scss|less|sass)$/);
-global.registerWebpackModules(appCssContext);
-application.loadAppCss();
-
-require("bundle-entry-points");
